v2.0
Added the possibility to save the density field.
The binary density field still has a bug, but one can save the density field as a txt file.
One problem was that in order to save the binary file, the mesh changes.
Another problem was that in order to select the halos, the mesh also changes. Thus, if one wants to save the density, one cannot save the density and halos at the same time.